finish syncTrades

+++ src/brokers/mylocalbroker/mylocalbroker.h
@@ -37,7 +37,7 @@ public:
 	virtual IBrokerControl::AllWallets getWallet() override;
 	virtual IStockApi::Ticker getTicker(const std::string_view &piar) override;
 	virtual json::Value getApiKeyFields() const override;
-	virtual long long iso8601ToMillis(const std::string &timeStr);
+	virtual unsigned long iso8601ToMillis(const std::string &timeStr);
 
 
 protected:
